src/main/java/com/my/controllers/LangServlet.java
package com.my.controllers;

import javax.servlet.ServletException;
import javax.servlet.annotation.WebServlet;
import javax.servlet.http.HttpServlet;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import javax.servlet.http.HttpSession;
import java.io.IOException;

/**
 * Controller for Changing language
 */

@WebServlet("/changeLang")
public class LangServlet extends HttpServlet {
    @Override
    public void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
        HttpSession session = req.getSession();
        String lang = req.getParameter("lang");
        String url = req.getHeader("Referer");
        session.setAttribute("lang", lang);
        resp.sendRedirect(url);
    }
}
